problem fixed

+++ src/ChallangeSys.java
@@ -1,7 +1,7 @@
 
 public class ChallangeSys {
 
-	public static boolean isPrime(int num) {
+	public static boolean isPrime(double num) {
 		
 		for(int i=2;i<num;i++) {
 			if(num%i==0)
